app - extract app logic to usecases sub module
